requirements.txt
aiolimiter==1.2.1
anyio==4.8.0
APScheduler==3.11.0
asgiref==3.8.1
cachetools==5.5.1
certifi==2024.12.14
charset-normalizer==3.4.1
colorama==0.4.6
Django==5.1.5
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
iniconfig==2.0.0
packaging==24.2
pluggy==1.5.0
pytest==8.3.4
python-telegram-bot==21.10
requests==2.32.3
sniffio==1.3.1
sqlparse==0.5.3
tornado==6.4.2
typing_extensions==4.12.2
tzdata==2025.1
tzlocal==5.2
urllib3==2.3.0
